Time: 0 ms (100%), Space: 7.7 MB (83.03%) - LeetHub

+++ 0342-power-of-four/0342-power-of-four.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPowerOfFour(int n) {
         int x = 0;
-        long long num = pow(4,0);
+        long num = pow(4,0);
 
         while(num<=n){
             if(num == n) return true;
